+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <!--
 ## Unreleased
+- Importing assessments with commas in inflection values will now be caught with a helpful error message
+=======
 ### Fixed
 - ContentPage import: Accept xlsx where field formatting is numeric.
 -->